refactor: Modularize socket configuration
- integrate`jwt` verification and change `configureSocket` parameter name from `server` to `io` for clarity in `socketIo.js`.
- implement `handleErrorEvents` helper function in `errorController.js`.
- move socket events `message friend`, `message room`, `chat activity` to `handleMessages` in `messageController.js`.
- move socket events `joinRoom` and `leaveRoom` to `handleRoomEvents` in `roomController.js`.
- integreate socket handlers `handleMessages`, `handleRoomEvents` and `handleErrorEvents` to `manageSocketEvents` in `socketController.js`.
- integrate `manageSocketEvents` with `ioServer` in `server.js`.
