+++ backend/controller/socket/errorController.js
@@ -1,0 +1,11 @@
+const handleErrorEvents = (socket) => {
+  socket.on("connect_error", (err) => {
+    console.error("Socket connection error:", err.message);
+  });
+
+  socket.on("error", (err) => {
+    console.error("Socket error:", err.message);
+  });
+};
+
+export default handleErrorEvents;
